Add support for quiet output of otp code command
When using this option the output of `pass otp -q ...` is predictable and can be
further used without parsing even when the password store is a repository as
well.  This is useful, for example, when using rofi-pass.

The patch is written in a way that other commands can optionally use the quiet
option for otp_insert as well.